Navbar component almost done
