+++ src/pages/home/home.tsx
@@ -1,40 +1,14 @@
-import { Container, Header, Links, MenuButton, Navbar, Perfil, TextBox } from "./styles"
+import { Container, Perfil, TextBox } from "./styles"
 import PerfilImg from "../../imgs/perfil.png"
 import { Content } from "../../components/Content/styles"
 import { TypeAnimation } from "react-type-animation"
-import { useState } from "react"
-
+import { Navbar } from "../../components/Navbar"
 
 export const Home = () => {
-    
-    const [navbar, setNavbar] = useState(true)
-
-    const showNavbar = () => setNavbar(!navbar)
 
     return (
         <>
-            <Header>
-                <Navbar>
-                    <div className="logo">PZE</div>
-                    <Links className="li">
-                        <li><a href="#Home">Home</a></li>
-                        <li><a href="#Skill">Skills</a></li>
-                        <li><a href="#Trajectory">Trajectory</a></li>
-                        <li><a href="#Contact">Contact</a></li>
-                    </Links>
-                    <MenuButton className="menu_btn" onClick={showNavbar}>
-                        <i className={navbar ? "fa-solid fa-bars" : "fa-solid fa-xmark"}></i>
-                        {/* <i className="fa-solid fa-bars"></i> */}
-                    </MenuButton>
-                        <section className={navbar ? "hide" : "show"}>
-                            <li><a href="#Home">Home</a></li>
-                            <li><a href="#Skill">Skills</a></li>
-                            <li><a href="#Trajectory">Trajectory</a></li>
-                            <li><a href="#Contact">Contact</a></li>
-                        </section>
-                </Navbar>
-            </Header>
-                
+            <Navbar />
             <Container id="Home">
                 <Content>
                     <TextBox>
